Move helper method to a new helpers file in channels folder
